cleanup: remove APIs we won't be exposing soon
This removes the APIs that we don't intend to expose in the beginning.

Leaving only the minimum necessary things for folks to self-enroll
and for mediator to be useful.

This also removes the tests that are no longer relevant.

Closes: https://github.com/stacklok/mediator/issues/1141
