+++ internal/controlplane/handlers_user_test.go
@@ -27,7 +27,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
-	"google.golang.org/protobuf/types/known/timestamppb"
 
 	mockdb "github.com/stacklok/mediator/database/mock"
 	"github.com/stacklok/mediator/internal/auth"
@@ -418,378 +417,6 @@ func TestDeleteUser_gRPC(t *testing.T) {
 			require.NoError(t, err, "failed to create test server")
 
 			resp, err := server.DeleteUser(ctx, tc.req)
-			tc.checkResponse(t, resp, err)
-		})
-	}
-}
-
-func TestGetUsersDBMock(t *testing.T) {
-	t.Parallel()
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockStore := mockdb.NewMockStore(ctrl)
-
-	orgID := uuid.New()
-	projectID := uuid.New()
-
-	request := &pb.GetUsersRequest{}
-
-	expectedUsers := []db.User{
-		{
-			ID:        1,
-			CreatedAt: time.Now(),
-			UpdatedAt: time.Now(),
-		},
-		{
-			ID:        2,
-			FirstName: sql.NullString{String: "Foo", Valid: true},
-			Email:     sql.NullString{String: "test@stacklok.com", Valid: true},
-			CreatedAt: time.Now(),
-			UpdatedAt: time.Now(),
-		},
-	}
-	// Create a new context and set the claims value
-	ctx := auth.WithPermissionsContext(context.Background(), auth.UserPermissions{
-		UserId:         1,
-		OrganizationId: orgID,
-		ProjectIds:     []uuid.UUID{projectID},
-		IsStaff:        true, // TODO: remove this
-		Roles: []auth.RoleInfo{
-			{RoleID: 1, IsAdmin: true, ProjectID: &projectID, OrganizationID: orgID}},
-	})
-
-	mockStore.EXPECT().ListUsers(ctx, gomock.Any()).
-		Return(expectedUsers, nil)
-
-	crypeng := crypto.NewEngine("test")
-
-	server := &Server{
-		store: mockStore,
-		cfg: &config.Config{
-			Salt: config.DefaultConfigForTest().Salt,
-		},
-		cryptoEngine: crypeng,
-	}
-
-	response, err := server.GetUsers(ctx, request)
-
-	assert.NoError(t, err)
-	assert.NotNil(t, response)
-	assert.Equal(t, len(expectedUsers), len(response.Users))
-	assert.Equal(t, expectedUsers[0].ID, response.Users[0].Id)
-	assert.Equal(t, expectedUsers[0].OrganizationID.String(), response.Users[0].OrganizationId)
-	expectedCreatedAt := expectedUsers[0].CreatedAt.In(time.UTC)
-	assert.Equal(t, expectedCreatedAt, response.Users[0].CreatedAt.AsTime().In(time.UTC))
-	expectedUpdatedAt := expectedUsers[0].UpdatedAt.In(time.UTC)
-	assert.Equal(t, expectedUpdatedAt, response.Users[0].UpdatedAt.AsTime().In(time.UTC))
-}
-
-func TestGetUsers_gRPC(t *testing.T) {
-	t.Parallel()
-
-	orgID := uuid.New()
-	projectID := uuid.New()
-
-	testCases := []struct {
-		name               string
-		req                *pb.GetUsersRequest
-		buildStubs         func(store *mockdb.MockStore)
-		checkResponse      func(t *testing.T, res *pb.GetUsersResponse, err error)
-		expectedStatusCode codes.Code
-	}{
-		{
-			name: "Success",
-			req:  &pb.GetUsersRequest{},
-			buildStubs: func(store *mockdb.MockStore) {
-				store.EXPECT().ListUsers(gomock.Any(), gomock.Any()).
-					Return([]db.User{
-						{
-							ID:              1,
-							IdentitySubject: "subject1",
-							OrganizationID:  orgID,
-							CreatedAt:       time.Now(),
-							UpdatedAt:       time.Now(),
-						},
-						{
-							ID:              2,
-							IdentitySubject: "subject2",
-							OrganizationID:  orgID,
-							Email:           sql.NullString{String: "test1@stacklok.com", Valid: true},
-							FirstName:       sql.NullString{String: "Foo", Valid: true},
-							CreatedAt:       time.Now(),
-							UpdatedAt:       time.Now(),
-						},
-					}, nil).
-					Times(1)
-			},
-			checkResponse: func(t *testing.T, res *pb.GetUsersResponse, err error) {
-				t.Helper()
-
-				firstNamePtr := "Foo"
-				emailPtr := "test1@stacklok.com"
-				expectedOrgs := []*pb.UserRecord{
-					{
-						Id:              1,
-						IdentitySubject: "subject1",
-						OrganizationId:  orgID.String(),
-						CreatedAt:       timestamppb.New(time.Now()),
-						UpdatedAt:       timestamppb.New(time.Now()),
-					},
-					{
-						Id:              2,
-						IdentitySubject: "subject2",
-						OrganizationId:  orgID.String(),
-						FirstName:       &firstNamePtr,
-						Email:           &emailPtr,
-						CreatedAt:       timestamppb.New(time.Now()),
-						UpdatedAt:       timestamppb.New(time.Now()),
-					},
-				}
-
-				assert.NoError(t, err)
-				assert.NotNil(t, res)
-				assert.Equal(t, len(expectedOrgs), len(res.Users))
-				assert.Equal(t, expectedOrgs[1].Id, res.Users[1].Id)
-				assert.Equal(t, expectedOrgs[1].OrganizationId, res.Users[1].OrganizationId)
-				assert.Equal(t, expectedOrgs[1].FirstName, res.Users[1].FirstName)
-				assert.Equal(t, *expectedOrgs[1].Email, *res.Users[1].Email)
-			},
-			expectedStatusCode: codes.OK,
-		},
-	}
-	// Create a new context and set the claims value
-	ctx := auth.WithPermissionsContext(context.Background(), auth.UserPermissions{
-		UserId:         1,
-		OrganizationId: orgID,
-		ProjectIds:     []uuid.UUID{projectID},
-		Roles: []auth.RoleInfo{
-			{RoleID: 1, IsAdmin: true, ProjectID: &projectID, OrganizationID: orgID}},
-	})
-
-	for i := range testCases {
-		tc := testCases[i]
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-
-			mockStore := mockdb.NewMockStore(ctrl)
-			tc.buildStubs(mockStore)
-			mockJwtValidator := mockjwt.NewMockJwtValidator(ctrl)
-
-			evt, err := events.Setup()
-			require.NoError(t, err, "failed to setup eventer")
-			server, err := NewServer(mockStore, evt, &config.Config{
-				Salt: config.DefaultConfigForTest().Salt,
-				Auth: config.AuthConfig{
-					TokenKey: generateTokenKey(t),
-				},
-			}, mockJwtValidator)
-			require.NoError(t, err, "failed to create test server")
-
-			resp, err := server.GetUsers(ctx, tc.req)
-			tc.checkResponse(t, resp, err)
-		})
-	}
-}
-
-func TestGetUserDBMock(t *testing.T) {
-	t.Parallel()
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockStore := mockdb.NewMockStore(ctrl)
-
-	request := &pb.GetUserByIdRequest{UserId: 1}
-
-	orgID := uuid.New()
-	projectID := uuid.New()
-
-	expectedUser := db.User{
-		ID:             1,
-		OrganizationID: orgID,
-		CreatedAt:      time.Now(),
-		UpdatedAt:      time.Now(),
-	}
-	// Create a new context and set the claims value
-	ctx := auth.WithPermissionsContext(context.Background(), auth.UserPermissions{
-		UserId:         1,
-		OrganizationId: orgID,
-		ProjectIds:     []uuid.UUID{projectID},
-		Roles: []auth.RoleInfo{
-			{RoleID: 1, IsAdmin: true, ProjectID: &projectID, OrganizationID: orgID}},
-	})
-
-	mockStore.EXPECT().GetUserByID(ctx, gomock.Any()).
-		Return(expectedUser, nil)
-	mockStore.EXPECT().GetUserRoles(ctx, gomock.Any())
-	mockStore.EXPECT().GetUserProjects(ctx, gomock.Any())
-
-	crypeng := crypto.NewEngine("test")
-
-	server := &Server{
-		store: mockStore,
-		cfg: &config.Config{
-			Salt: config.DefaultConfigForTest().Salt,
-		},
-		cryptoEngine: crypeng,
-	}
-
-	response, err := server.GetUserById(ctx, request)
-
-	assert.NoError(t, err)
-	assert.NotNil(t, response)
-	assert.Equal(t, expectedUser.ID, response.User.Id)
-	assert.Equal(t, expectedUser.OrganizationID.String(), response.User.OrganizationId)
-	expectedCreatedAt := expectedUser.CreatedAt.In(time.UTC)
-	assert.Equal(t, expectedCreatedAt, response.User.CreatedAt.AsTime().In(time.UTC))
-	expectedUpdatedAt := expectedUser.UpdatedAt.In(time.UTC)
-	assert.Equal(t, expectedUpdatedAt, response.User.UpdatedAt.AsTime().In(time.UTC))
-}
-
-func TestGetNonExistingUserDBMock(t *testing.T) {
-	t.Parallel()
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	orgID := uuid.New()
-	projectID := uuid.New()
-
-	mockStore := mockdb.NewMockStore(ctrl)
-
-	request := &pb.GetUserByIdRequest{UserId: 5}
-	// Create a new context and set the claims value
-	ctx := auth.WithPermissionsContext(context.Background(), auth.UserPermissions{
-		UserId:         1,
-		OrganizationId: orgID,
-		ProjectIds:     []uuid.UUID{projectID},
-		IsStaff:        true, // TODO: remove this
-		Roles: []auth.RoleInfo{
-			{RoleID: 1, IsAdmin: true, ProjectID: &projectID, OrganizationID: orgID}},
-	})
-
-	mockStore.EXPECT().GetUserByID(ctx, gomock.Any()).
-		Return(db.User{}, sql.ErrNoRows)
-
-	crypeng := crypto.NewEngine("test")
-
-	server := &Server{
-		store: mockStore,
-		cfg: &config.Config{
-			Salt: config.DefaultConfigForTest().Salt,
-		},
-		cryptoEngine: crypeng,
-	}
-
-	response, err := server.GetUserById(ctx, request)
-
-	assert.Error(t, err)
-	assert.Nil(t, response)
-}
-
-func TestGetUser_gRPC(t *testing.T) {
-	t.Parallel()
-
-	orgID := uuid.New()
-	projectID := uuid.New()
-
-	testCases := []struct {
-		name               string
-		req                *pb.GetUserByIdRequest
-		buildStubs         func(store *mockdb.MockStore)
-		checkResponse      func(t *testing.T, res *pb.GetUserByIdResponse, err error)
-		expectedStatusCode codes.Code
-	}{
-		{
-			name: "Success",
-			req:  &pb.GetUserByIdRequest{UserId: 1},
-			buildStubs: func(store *mockdb.MockStore) {
-				store.EXPECT().GetUserByID(gomock.Any(), gomock.Any()).
-					Return(db.User{
-						ID:             1,
-						OrganizationID: orgID,
-						CreatedAt:      time.Now(),
-						UpdatedAt:      time.Now(),
-					}, nil).
-					Times(1)
-				store.EXPECT().GetUserRoles(gomock.Any(), gomock.Any())
-				store.EXPECT().GetUserProjects(gomock.Any(), gomock.Any())
-			},
-			checkResponse: func(t *testing.T, res *pb.GetUserByIdResponse, err error) {
-				t.Helper()
-
-				expectedUser := pb.UserRecord{
-					Id:             1,
-					OrganizationId: orgID.String(),
-					CreatedAt:      timestamppb.New(time.Now()),
-					UpdatedAt:      timestamppb.New(time.Now()),
-				}
-
-				assert.NoError(t, err)
-				assert.NotNil(t, res)
-				assert.Equal(t, expectedUser.Id, res.User.Id)
-				assert.Equal(t, expectedUser.OrganizationId, res.User.OrganizationId)
-			},
-			expectedStatusCode: codes.OK,
-		},
-		{
-			name: "NonExisting",
-			req:  &pb.GetUserByIdRequest{UserId: 5},
-			buildStubs: func(store *mockdb.MockStore) {
-				store.EXPECT().GetUserByID(gomock.Any(), gomock.Any()).
-					Return(db.User{}, nil).
-					Times(1)
-				store.EXPECT().GetUserRoles(gomock.Any(), gomock.Any())
-				store.EXPECT().GetUserProjects(gomock.Any(), gomock.Any())
-
-			},
-			checkResponse: func(t *testing.T, res *pb.GetUserByIdResponse, err error) {
-				t.Helper()
-
-				assert.NoError(t, err)
-				assert.Equal(t, int32(0), res.User.Id)
-			},
-			expectedStatusCode: codes.OK,
-		},
-	}
-	// Create a new context and set the claims value
-	ctx := auth.WithPermissionsContext(context.Background(), auth.UserPermissions{
-		UserId:         1,
-		OrganizationId: orgID,
-		ProjectIds:     []uuid.UUID{projectID},
-		IsStaff:        true, // TODO: remove this
-		Roles: []auth.RoleInfo{
-			{RoleID: 1, IsAdmin: true, ProjectID: &projectID, OrganizationID: orgID}},
-	})
-
-	for i := range testCases {
-		tc := testCases[i]
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-
-			mockStore := mockdb.NewMockStore(ctrl)
-			tc.buildStubs(mockStore)
-			mockJwtValidator := mockjwt.NewMockJwtValidator(ctrl)
-
-			evt, err := events.Setup()
-			require.NoError(t, err, "failed to setup eventer")
-			server, err := NewServer(mockStore, evt, &config.Config{
-				Salt: config.DefaultConfigForTest().Salt,
-				Auth: config.AuthConfig{
-					TokenKey: generateTokenKey(t),
-				},
-			}, mockJwtValidator)
-			require.NoError(t, err, "failed to create test server")
-
-			resp, err := server.GetUserById(ctx, tc.req)
 			tc.checkResponse(t, resp, err)
 		})
 	}
